refactor(realm): extract sliccy:agent module from js-realm-shared
